feat: enhance CLI, fix bugs, improve signatures and confidence algorithm
- Fix recover hardcoded LBA 0 bug (use actual partition start)
- Fix RIFF signature ambiguity (WAV/AVI/WebP sub-type detection)
- Fix all MSVC warnings (type conversions, strcasecmp compat)
- Add --type, --min-size, --json, --verbose to scan command
- Add --force, --verbose, scan progress to recover command
- Add batch recovery (multiple IDs: 1,2,3 or range: 1-10)
- Add result filter functions (by category and min size)
- Add stdout JSON output for scan results
- Improve confidence algorithm with multi-factor scoring
- Add test_scanner.c with 12 new test cases
- Bump version to 0.2.0

+++ src/fs/ntfs/data_run.c
@@ -165,7 +165,7 @@ int ntfs_read_file_data(ntfs_volume_t *vol, const ntfs_file_info_t *info,
                 return SALVAGE_ERR_NO_MEMORY;
             }
             
-            ret = device_read_sectors(vol->device, read_lba, read_sectors, tmp);
+            ret = device_read_sectors(vol->device, read_lba, (uint32_t)read_sectors, tmp);
             if (ret != SALVAGE_OK) {
                 free(tmp);
                 free(record);
